requirements.txt
langchain
# Core LLM + Chains
langchain
langchain-openai
openai

# LangGraph
langgraph

# RAG + Semantic Search
chromadb
langchain-chroma
tiktoken
sentence-transformers

# Environment
python-dotenv

# Optional but useful for your deliverables
python-docx
reportlab
presidio-analyzer
presidio-anonymizer
